add guard and decorator for subscriptions
