+++ apps/api/src/modules/quiz/controller/list.quiz.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, Inject, Post, Res, UnauthorizedException } from '@nestjs/common';
+import { Body, Controller, Get, Inject, Post, Res, UnauthorizedException, UseGuards } from '@nestjs/common';
 
 import {
   TYPES,
@@ -8,6 +8,9 @@ import { Role } from 'src/modules/user/domain/role.enum';
 import { AuthController } from 'src/utils/decorators/auth-controller.decorator';
 import { IListQuizService } from '../interfaces/services/list.quiz.service.interface';
 import { SpaceId } from 'src/modules/auth/decorators';
+import { SubscriptionGuard } from 'src/modules/subscription/guards/subscription.guard';
+import { SubscriptionDecorator } from 'src/modules/subscription/decorators/subscription.decorator';
+import { CachedSubscription } from 'src/modules/subscription/dto/cached-response.dto';
 
 @AuthController('quiz')
 export class ListQuizController {
@@ -18,10 +21,13 @@ export class ListQuizController {
 
   @Get()
   @Roles(Role.SpaceAdmin)
+  @UseGuards(SubscriptionGuard)
   async list(
-    @SpaceId() spaceId: number
+    @SpaceId() spaceId: number,
+    @SubscriptionDecorator() subscription: Partial<CachedSubscription>
   ) 
   { 
+    console.log("🚀 ~ ListQuizController ~ list ~ subscription:", subscription)
     const quizzes = await this.listQuizService.execute(spaceId)
     return quizzes
   }
