+++ apps/api/src/modules/user/controllers/me.user.controller.ts
@@ -17,8 +17,7 @@ export class MeUserController {
 
   @Get()
   @Roles(Role.SpaceAdmin)
-  async me(@LoggedUser() user: LoggedUserDto) {
-    console.log("🚀 ~ MeUserController ~ me ~ user:", user)
+  async me(@LoggedUser() user: LoggedUserDto) {    
     await this.markUserLoginService.execute(user.id);
     return user
   }
